Refactor Animated[List, Grid, SliverList, SliverGrid] to share common code (#113793)
