+++ packages/flutter/lib/widgets.dart
@@ -19,8 +19,7 @@ export 'foundation.dart' show UniqueKey;
 export 'rendering.dart' show TextSelectionHandleType;
 export 'src/widgets/actions.dart';
 export 'src/widgets/animated_cross_fade.dart';
-export 'src/widgets/animated_grid.dart';
-export 'src/widgets/animated_list.dart';
+export 'src/widgets/animated_scroll_view.dart';
 export 'src/widgets/animated_size.dart';
 export 'src/widgets/animated_switcher.dart';
 export 'src/widgets/annotated_region.dart';
